+++ packages/expo-router/src/native-tabs/__tests__/options.test.ios.tsx
@@ -41,8 +41,8 @@ it('can pass options via options prop', () => {
   renderRouter({
     _layout: () => (
       <NativeTabs>
-        <NativeTabs.Trigger name="index" options={{ ...indexOptions }} />
-        <NativeTabs.Trigger name="second" options={{ ...secondOptions }} />
+        <NativeTabs.Trigger name="index" unstable_options={{ ...indexOptions }} />
+        <NativeTabs.Trigger name="second" unstable_options={{ ...secondOptions }} />
       </NativeTabs>
     ),
     index: () => <View testID="index" />,
@@ -655,12 +655,12 @@ describe('Dynamic options', () => {
     renderRouter({
       _layout: () => (
         <NativeTabs>
-          <NativeTabs.Trigger name="index" options={{ title: 'Initial Title' }} />
+          <NativeTabs.Trigger name="index" unstable_options={{ title: 'Initial Title' }} />
         </NativeTabs>
       ),
       index: () => (
         <View testID="index">
-          <NativeTabs.Trigger name="index" options={{ title: 'Updated Title' }} />
+          <NativeTabs.Trigger name="index" unstable_options={{ title: 'Updated Title' }} />
         </View>
       ),
     });
@@ -676,7 +676,7 @@ describe('Dynamic options', () => {
     renderRouter({
       _layout: () => (
         <NativeTabs>
-          <NativeTabs.Trigger name="index" options={{ title: 'Initial Title' }} />
+          <NativeTabs.Trigger name="index" unstable_options={{ title: 'Initial Title' }} />
         </NativeTabs>
       ),
       index: () => (
@@ -721,7 +721,7 @@ describe('Dynamic options', () => {
         <View testID="index">
           <NativeTabs.Trigger
             name="index"
-            options={{
+            unstable_options={{
               title: 'Updated Title',
               badgeValue: '5',
               icon: { sf: 'homepod.2.fill' },
@@ -825,8 +825,8 @@ describe('Dynamic options', () => {
     renderRouter({
       _layout: () => (
         <NativeTabs>
-          <NativeTabs.Trigger name="index" options={{ title: 'Initial Title' }} />
-          <NativeTabs.Trigger name="second" options={{ title: 'Second' }} />
+          <NativeTabs.Trigger name="index" unstable_options={{ title: 'Initial Title' }} />
+          <NativeTabs.Trigger name="second" unstable_options={{ title: 'Second' }} />
         </NativeTabs>
       ),
       index: () => (